+++ container-engine-lib/lib/backend_impls/kubernetes/kubernetes_kurtosis_backend_user_service_functions.go
@@ -2,7 +2,6 @@ package kubernetes
 import (
 	"bytes"
 	"context"
-	"fmt"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/kubernetes/kubernetes_resource_collectors"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/kubernetes/object_attributes_provider"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/kubernetes/object_attributes_provider/annotation_key_consts"
@@ -17,12 +16,13 @@ import (
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/files_artifact_expansion_volume"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/port_spec"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/service"
+	"github.com/kurtosis-tech/container-engine-lib/lib/uuid_generator"
 	"github.com/kurtosis-tech/stacktrace"
 	"github.com/sirupsen/logrus"
 	"io"
 	apiv1 "k8s.io/api/core/v1"
+	applyconfigurationsv1 "k8s.io/client-go/applyconfigurations/core/v1"
 	"net"
-	"time"
 )
 
 /*
@@ -65,11 +65,12 @@ const (
 	// Our user services don't need service accounts
 	userServiceServiceAccountName = ""
 
-	// Kubernetes doesn't allow us to create services without ports exposed, but we won't have the ports that the user
-	// wants to listen to until they StartService. We create the Kubernetes Service on RegisterService, so we need to
-	// set up a notional port until the user calls StartService.
-	unboundPortName = "nonexistent"
-	unboundPortNumber = 80
+	// Kubernetes doesn't allow us to create services without ports exposed, but we might not have ports in the following situations:
+	//  1) we've registered a service but haven't started a container yet (so ports are yet to come)
+	//  2) we've started a container that doesn't listen on any ports
+	// In these cases, we use these notional unbound ports
+	unboundPortName = "nonexistent-port"
+	unboundPortNumber = 1
 )
 
 // Kubernetes doesn't provide public IP or port information; this is instead handled by the Kurtosis gateway that the user uses
@@ -104,12 +105,11 @@ func (backend *KubernetesKurtosisBackend) RegisterUserService(ctx context.Contex
 		return nil, stacktrace.Propagate(err, "An error occurred getting namespace name for enclave '%v'", enclaveId)
 	}
 
-	// TODO Switch this, and all other GUIDs, to a UUID??
-	serviceGuid := service.ServiceGUID(fmt.Sprintf(
-		"%v-%v",
-		serviceId,
-		time.Now().Unix(),
-	))
+	serviceGuidStr, err := uuid_generator.GenerateUUIDString()
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "An error occurred generating a UUID to use for the service GUID")
+	}
+	serviceGuid := service.ServiceGUID(serviceGuidStr)
 
 	objectAttributesProvider := object_attributes_provider.GetKubernetesObjectAttributesProvider()
 	enclaveObjAttributesProvider := objectAttributesProvider.ForEnclave(enclaveId)
@@ -228,7 +228,7 @@ func (backend *KubernetesKurtosisBackend) StartUserService(
 		return nil, stacktrace.Propagate(err, "An error occurred getting user service objects and Kubernetes resources matching service GUID '%v'", serviceGuid)
 	}
 	if len(preexistingObjectsAndResources) == 0 {
-		return nil, stacktrace.NewError("Couldn't find any services registrations matching service GUID '%v'", serviceGuid)
+		return nil, stacktrace.NewError("Couldn't find any service registrations matching service GUID '%v'", serviceGuid)
 	}
 	if len(preexistingObjectsAndResources) > 1 {
 		// Should never happen because service GUIDs should be unique
@@ -308,33 +308,11 @@ func (backend *KubernetesKurtosisBackend) StartUserService(
 		}
 	}()
 
-	// Update the service to:
-	// - Set the service ports appropriately
-	// - Irrevocably record that a pod is bound to the service (so that even if the pod is deleted, the service won't
-	// 	 be usable again
-	serializedPortSpecs, err := kubernetes_port_spec_serializer.SerializePortSpecs(privatePorts)
+
+	updatedService, err := backend.updateServiceWhenContainerStarted(ctx, namespaceName, kubernetesService, privatePorts)
 	if err != nil {
-		return nil, stacktrace.Propagate(err, "An error occurred serializing the following private port specs: %+v", privatePorts)
+		return nil, stacktrace.Propagate(err, "An error occurred updating service '%v' to reflect its new ports: %+v", kubernetesService.GetName(), privatePorts)
 	}
-	kubernetesServicePorts, err := getKubernetesServicePortsFromPrivatePortSpecs(privatePorts)
-	if err != nil {
-		return nil, stacktrace.Propagate(err, "An error occurred getting Kubernetes service ports for the following private port specs: %+v", privatePorts)
-	}
-	updatedService := kubernetesService.DeepCopy()
-	updatedService.Annotations[annotation_key_consts.PortSpecsAnnotationKey.GetString()] = serializedPortSpecs.GetString()
-	updatedService.Spec.Ports = kubernetesServicePorts
-	if err := backend.kubernetesManager.UpdateService(ctx, namespaceName, updatedService); err != nil {
-		return nil, stacktrace.Propagate(err, "An error occurred updating the user service to open the service ports and add the serialized private port specs annotation")
-	}
-	shouldUndoServiceUpdate := true
-	defer func() {
-		if shouldUndoServiceUpdate {
-			if err := backend.kubernetesManager.UpdateService(ctx, namespaceName, kubernetesService); err != nil {
-				logrus.Errorf("Starting service didn't complete successfully so we tried to undo the changes we made to the Kubernetes service but doing so threw an error:\n%v", err)
-				logrus.Errorf("ACTION REQUIRED: You'll need to manually correct service '%v' in namespace '%v'!!!", kubernetesService.GetName(), namespaceName)
-			}
-		}
-	}()
 
 	kubernetesResources := map[service.ServiceGUID]*userServiceKubernetesResources{
 		serviceGuid: {
@@ -356,7 +334,6 @@ func (backend *KubernetesKurtosisBackend) StartUserService(
 	}
 
 	shouldDestroyPod = false
-	shouldUndoServiceUpdate = false
 	return objectsAndResources.service, nil
 }
 
@@ -365,8 +342,25 @@ func (backend *KubernetesKurtosisBackend) GetUserServices(
 	enclaveId enclave.EnclaveID,
 	filters *service.ServiceFilters,
 ) (successfulUserServices map[service.ServiceGUID]*service.Service, resultError error) {
-	//TODO implement me
-	panic("implement me")
+	allObjectsAndResources, err := backend.getMatchingUserServiceObjectsAndKubernetesResources(ctx, enclaveId, filters)
+	if err != nil {
+		return nil, stacktrace.Propagate(
+			err,
+			"An error occurred getting user services in enclave '%v' matching filters: %+v",
+			enclaveId,
+			filters,
+		)
+	}
+	result := map[service.ServiceGUID]*service.Service{}
+	for guid, serviceObjsAndResources := range allObjectsAndResources {
+		serviceObj := serviceObjsAndResources.service
+		if serviceObj == nil {
+			// Indicates a registration-only service; skip
+			continue
+		}
+		result[guid] = serviceObj
+	}
+	return result, nil
 }
 
 func (backend *KubernetesKurtosisBackend) GetUserServiceLogs(
@@ -510,19 +504,106 @@ func (backend *KubernetesKurtosisBackend) CopyFromUserService(ctx context.Contex
 	panic("implement me")
 }
 
-func (backend *KubernetesKurtosisBackend) StopUserServices(ctx context.Context, enclaveId enclave.EnclaveID, filters *service.ServiceFilters) (successfulUserServiceGuids map[service.ServiceGUID]bool, erroredUserServiceGuids map[service.ServiceGUID]error, resultErr error) {
-	// TODO kill the pod
-	// TODO remove the service's selectors
+func (backend *KubernetesKurtosisBackend) StopUserServices(ctx context.Context, enclaveId enclave.EnclaveID, filters *service.ServiceFilters) (resultSuccessfulGuids map[service.ServiceGUID]bool, resultErroredGuids map[service.ServiceGUID]error, resultErr error) {
+	namespaceName, err := backend.getEnclaveNamespaceName(ctx, enclaveId)
+	if err != nil {
+		return nil, nil, stacktrace.Propagate(err, "An error occurred getting namespace name for enclave '%v'", enclaveId)
+	}
 
-	//TODO implement me
-	panic("implement me")
+	allObjectsAndResources, err := backend.getMatchingUserServiceObjectsAndKubernetesResources(ctx, enclaveId, filters)
+	if err != nil {
+		return nil, nil, stacktrace.Propagate(err, "An error occurred getting user services in enclave '%v' matching filters: %+v", enclaveId, filters)
+	}
+
+	successfulGuids := map[service.ServiceGUID]bool{}
+	erroredGuids := map[service.ServiceGUID]error{}
+	for serviceGuid, serviceObjsAndResources := range allObjectsAndResources {
+		resources := serviceObjsAndResources.kubernetesResources
+
+		pod := resources.pod
+		if pod != nil {
+			if err := backend.kubernetesManager.RemovePod(ctx, namespaceName, pod.Name); err != nil {
+				erroredGuids[serviceGuid] = stacktrace.Propagate(
+					err,
+					"An error occurred removing Kubernetes pod '%v' in namespace '%v'",
+					pod.Name,
+					namespaceName,
+				)
+				continue
+			}
+		}
+
+		kubernetesService := resources.service
+		serviceName := kubernetesService.Name
+		updateConfigurator := func(updatesToApply *applyconfigurationsv1.ServiceApplyConfiguration) {
+			specUpdates := applyconfigurationsv1.ServiceSpec().WithSelector(nil)
+			updatesToApply.WithSpec(specUpdates)
+		}
+		if _, err := backend.kubernetesManager.UpdateService(ctx, namespaceName, serviceName, updateConfigurator); err != nil {
+			erroredGuids[serviceGuid] = stacktrace.Propagate(
+				err,
+				"An error occurred updating service '%v' in namespace '%v' to reflect that it's no longer running",
+				serviceName,
+				namespaceName,
+			)
+			continue
+		}
+
+		successfulGuids[serviceGuid] = true
+	}
+	return successfulGuids, erroredGuids, nil
 }
 
-func (backend *KubernetesKurtosisBackend) DestroyUserServices(ctx context.Context, enclaveId enclave.EnclaveID, filters *service.ServiceFilters) (successfulUserServiceGuids map[service.ServiceGUID]bool, erroredUserServiceGuids map[service.ServiceGUID]error, resultErr error) {
-	// TODO Destroy persistent volume claims (???)
+func (backend *KubernetesKurtosisBackend) DestroyUserServices(ctx context.Context, enclaveId enclave.EnclaveID, filters *service.ServiceFilters) (resultSuccessfulGuids map[service.ServiceGUID]bool, resultErroredGuids map[service.ServiceGUID]error, resultErr error) {
+	namespaceName, err := backend.getEnclaveNamespaceName(ctx, enclaveId)
+	if err != nil {
+		return nil, nil, stacktrace.Propagate(err, "An error occurred getting namespace name for enclave '%v'", enclaveId)
+	}
 
-	//TODO implement me
-	panic("implement me")
+	allObjectsAndResources, err := backend.getMatchingUserServiceObjectsAndKubernetesResources(ctx, enclaveId, filters)
+	if err != nil {
+		return nil, nil, stacktrace.Propagate(
+			err,
+			"An error occurred getting user services in enclave '%v' matching filters: %+v",
+			enclaveId,
+			filters,
+		)
+	}
+
+	successfulGuids := map[service.ServiceGUID]bool{}
+	erroredGuids := map[service.ServiceGUID]error{}
+	for serviceGuid, serviceObjsAndResources := range allObjectsAndResources {
+		resources := serviceObjsAndResources.kubernetesResources
+
+		pod := resources.pod
+		if pod != nil {
+			if err := backend.kubernetesManager.RemovePod(ctx, namespaceName, pod.Name); err != nil {
+				erroredGuids[serviceGuid] = stacktrace.Propagate(
+					err,
+					"An error occurred removing Kubernetes pod '%v' in namespace '%v'",
+					pod.Name,
+					namespaceName,
+				)
+				continue
+			}
+		}
+
+		kubernetesService := resources.service
+		if kubernetesService != nil {
+			if err := backend.kubernetesManager.RemoveService(ctx, namespaceName, kubernetesService.Name); err != nil {
+				erroredGuids[serviceGuid] = stacktrace.Propagate(
+					err,
+					"An error occurred removing Kubernetes service '%v' in namespace '%v'",
+					kubernetesService.Name,
+					namespaceName,
+				)
+				continue
+			}
+		}
+
+		successfulGuids[serviceGuid] = true
+	}
+	return successfulGuids, erroredGuids, nil
 }
 
 
@@ -540,6 +621,10 @@ func (backend *KubernetesKurtosisBackend) getMatchingUserServiceObjectsAndKubern
 	allResources, err := backend.getUserServiceKubernetesResourcesMatchingGuids(ctx, enclaveId, filters.GUIDs)
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred getting user service Kubernetes resources matching GUIDs: %+v", filters.GUIDs)
+	}
+
+	for serviceGuid, serviceResources := range allResources {
+		logrus.Tracef("Found resources for service '%v': %+v", serviceGuid, serviceResources)
 	}
 
 	allObjectsAndResources, err := getUserServiceObjectsFromKubernetesResources(enclaveId, allResources)
@@ -654,6 +739,7 @@ func (backend *KubernetesKurtosisBackend) getUserServiceKubernetesResourcesMatch
 		return nil, stacktrace.Propagate(err, "An error occurred getting Kubernetes services matching service GUIDs: %+v", serviceGuids)
 	}
 	for serviceGuidStr, kubernetesServicesForGuid := range matchingKubernetesServices {
+		logrus.Tracef("Found Kubernetes services for GUID '%v': %+v", serviceGuidStr, kubernetesServicesForGuid)
 		serviceGuid := service.ServiceGUID(serviceGuidStr)
 
 		numServicesForGuid := len(kubernetesServicesForGuid)
@@ -671,6 +757,7 @@ func (backend *KubernetesKurtosisBackend) getUserServiceKubernetesResourcesMatch
 			resultObj = &userServiceKubernetesResources{}
 		}
 		resultObj.service = kubernetesService
+		results[serviceGuid] = resultObj
 	}
 
 	// Get k8s pods
@@ -686,6 +773,7 @@ func (backend *KubernetesKurtosisBackend) getUserServiceKubernetesResourcesMatch
 		return nil, stacktrace.Propagate(err, "An error occurred getting Kubernetes pods matching service GUIDs: %+v", serviceGuids)
 	}
 	for serviceGuidStr, kubernetesPodsForGuid := range matchingKubernetesPods {
+		logrus.Tracef("Found Kubernetes pods for GUID '%v': %+v", serviceGuidStr, kubernetesPodsForGuid)
 		serviceGuid := service.ServiceGUID(serviceGuidStr)
 
 		numPodsForGuid := len(kubernetesPodsForGuid)
@@ -703,6 +791,7 @@ func (backend *KubernetesKurtosisBackend) getUserServiceKubernetesResourcesMatch
 			resultObj = &userServiceKubernetesResources{}
 		}
 		resultObj.pod = kubernetesPod
+		results[serviceGuid] = resultObj
 	}
 
 	return results, nil
@@ -772,16 +861,9 @@ func getUserServiceObjectsFromKubernetesResources(
 			continue
 		}
 
-		podPhase := kubernetesPod.Status.Phase
-		isPodRunning, found := isPodRunningDeterminer[podPhase]
-		if !found {
-			// Should never happen because we enforce completeness of the determiner via unit test
-			return nil, stacktrace.NewError("No is-pod-running determination found for pod phase '%v' on pod '%v'; this is a bug in Kurtosis", podPhase, kubernetesPod.Name)
-		}
-		// TODO Rename this; this shouldn't be called "ContainerStatus" since there's no longer a 1:1 mapping between container:kurtosis_object
-		containerStatus := container_status.ContainerStatus_Stopped
-		if isPodRunning {
-			containerStatus = container_status.ContainerStatus_Running
+		containerStatus, err := getContainerStatusFromPod(resourcesToParse.pod)
+		if err != nil {
+			return nil, stacktrace.Propagate(err, "An error occurred getting container status from Kubernetes pod '%+v'", resourcesToParse.pod)
 		}
 
 		resultObj.service = service.NewService(
@@ -930,4 +1012,104 @@ func (backend *KubernetesKurtosisBackend) getUserServiceVolumeInfoFromFilesArtif
 	}
 
 	return podVolumes, containerMounts, nil
+}
+
+// Update the service to:
+// - Set the service ports appropriately
+// - Irrevocably record that a pod is bound to the service (so that even if the pod is deleted, the service won't
+// 	 be usable again
+func (backend *KubernetesKurtosisBackend) updateServiceWhenContainerStarted(
+	ctx context.Context,
+	namespaceName string,
+	kubernetesService *apiv1.Service,
+	privatePorts map[string]*port_spec.PortSpec,
+) (
+	*apiv1.Service,
+	error,
+) {
+	serviceName := kubernetesService.GetName()
+
+	serializedPortSpecs, err := kubernetes_port_spec_serializer.SerializePortSpecs(privatePorts)
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "An error occurred serializing the following private port specs: %+v", privatePorts)
+	}
+
+	// We only need to modify the ports from the default (unbound) ports if the user actually declares ports
+	newServicePorts := kubernetesService.Spec.Ports
+	if len(privatePorts) > 0 {
+		candidateNewServicePorts, err := getKubernetesServicePortsFromPrivatePortSpecs(privatePorts)
+		if err != nil {
+			return nil, stacktrace.Propagate(err, "An error occurred getting Kubernetes service ports for the following private port specs: %+v", privatePorts)
+		}
+		newServicePorts = candidateNewServicePorts
+	}
+
+	newAnnotations := kubernetesService.Annotations
+	if newAnnotations == nil {
+		newAnnotations = map[string]string{}
+	}
+	newAnnotations[annotation_key_consts.PortSpecsAnnotationKey.GetString()] = serializedPortSpecs.GetString()
+
+	updatingConfigurator := func(updatesToApply *applyconfigurationsv1.ServiceApplyConfiguration) {
+		specUpdateToApply := applyconfigurationsv1.ServiceSpec()
+		for _, newServicePort := range newServicePorts {
+			portUpdateToApply := &applyconfigurationsv1.ServicePortApplyConfiguration{
+				Name:        &newServicePort.Name,
+				Protocol:    &newServicePort.Protocol,
+				// TODO fill this out for an app port!
+				AppProtocol: nil,
+				Port:        &newServicePort.Port,
+			}
+			specUpdateToApply.WithPorts(portUpdateToApply)
+		}
+		updatesToApply.WithSpec(specUpdateToApply)
+
+		updatesToApply.WithAnnotations(newAnnotations)
+	}
+	updatedService, err := backend.kubernetesManager.UpdateService(ctx, namespaceName, serviceName, updatingConfigurator)
+	if err != nil {
+		return nil, stacktrace.Propagate(
+			err,
+			"An error occurred updating Kubernetes service '%v' in namespace '%v' to open the service ports and add " +
+				"the serialized private port specs annotation",
+			serviceName,
+			namespaceName,
+		)
+	}
+	shouldUndoUpdate := true
+	defer func() {
+		if shouldUndoUpdate {
+			undoingConfigurator := func(reversionToApply *applyconfigurationsv1.ServiceApplyConfiguration) {
+				specReversionToApply := applyconfigurationsv1.ServiceSpec()
+				for _, oldServicePort := range newServicePorts {
+					portUpdateToApply := &applyconfigurationsv1.ServicePortApplyConfiguration{
+						Name:        &oldServicePort.Name,
+						Protocol:    &oldServicePort.Protocol,
+						// TODO fill this out for an app port!
+						AppProtocol: nil,
+						Port:        &oldServicePort.Port,
+					}
+					specReversionToApply.WithPorts(portUpdateToApply)
+				}
+				reversionToApply.WithSpec(specReversionToApply)
+
+				reversionToApply.WithAnnotations(kubernetesService.Annotations)
+			}
+			if _, err := backend.kubernetesManager.UpdateService(ctx, namespaceName, serviceName, undoingConfigurator); err != nil {
+				logrus.Errorf(
+					"An error occurred updating Kubernetes service '%v' in namespace '%v' to open the service ports " +
+						"and add the serialized private port specs annotation so we tried to revert the service to " +
+						"its values before the update, but an error occurred; this means the service is likely in " +
+						"an inconsistent state:\n%v",
+					serviceName,
+					namespaceName,
+					err,
+				)
+				// Nothing we can really do here to recover
+			}
+		}
+	}()
+
+	shouldUndoUpdate = false
+	return updatedService, nil
 }